% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity.R
\name{aSiO2}
\alias{aSiO2}
\title{Final user-facing function !}
\usage{
aSiO2(Pbar, TK, muSiO2)
}
\arguments{
\item{Pbar}{Pressure, bars}

\item{TK}{Temperature, Kelvin}

\item{muSiO2}{Chemical potential of SiO2}
}
\value{
The activity of SiO2
}
\description{
Final user-facing function !
}
