% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity.R
\name{aTiO2}
\alias{aTiO2}
\title{Final user-facing function !}
\usage{
aTiO2(Pbar, TK, muTiO2)
}
\arguments{
\item{Pbar}{Pressure, bars}

\item{TK}{Temperature, Kelvin}

\item{muTiO2}{Chemical potential of TiO2}
}
\value{
The activity of TiO2
}
\description{
Final user-facing function !
}
