% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity.R
\name{CORK}
\alias{CORK}
\title{vectorized version of Gibbs
Compensated-Redlich-Kwong (CORK) Equation
Following Holland and Powell 1991
Jesse Walters
Rtlnf
Equations 9}
\usage{
CORK(Ca0, Ca1, Cb0, Cc0, Cc1, Cd0, Cd1, Tcf, Pcf, Pbar, TK)
}
\description{
vectorized version of Gibbs
Compensated-Redlich-Kwong (CORK) Equation
Following Holland and Powell 1991
Jesse Walters
Rtlnf
Equations 9
}
