% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity.R
\name{fO2}
\alias{fO2}
\title{Oxygen fugacity}
\usage{
fO2(Pbar, TK, muO2)

fO2(Pbar, TK, muO2)
}
\arguments{
\item{Pbar}{Pressure, bars}

\item{TK}{Temperature, Kelvin}

\item{muSiO2}{Chemical potential of SiO2}
}
\value{
a data frame with fO2 and delta FMQ
}
\description{
Oxygen fugacity

Final user-facing function !
}
