% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity.R
\name{GibbsPure}
\alias{GibbsPure}
\title{Molar Gibbs Free Energy Calculator for Pure Solid Phases
Jesse Walters
Gf
This function calculates the molar gibbs free energy for solid phases in
two parts: G=G0+RTlnK
where
G0=Hf-TS0-integral(Cp)dT-T*integral(Cp/T)dT+integral(Vsolid)dP}
\usage{
GibbsPure(
  TK,
  PGPa,
  Hff,
  S0f,
  af,
  bf,
  cf,
  df,
  Ca0,
  Ca1,
  Cb0,
  Cc0,
  Cc1,
  Cd0,
  Cd1,
  Tcf,
  Pcf
)
}
\description{
Molar Gibbs Free Energy Calculator for Pure Solid Phases
Jesse Walters
Gf
This function calculates the molar gibbs free energy for solid phases in
two parts: G=G0+RTlnK
where
G0=Hf-TS0-integral(Cp)dT-T*integral(Cp/T)dT+integral(Vsolid)dP
}
