% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity.R
\name{GibbsSolid_landau}
\alias{GibbsSolid_landau}
\title{This function calculates the molar gibbs free energy for solid phases in
two parts: G=G0+RTlnK
where
G0=Hf-TS0-integral(Cp)dT-T*integral(Cp/T)dT+integral(Vsolid)dP}
\usage{
GibbsSolid_landau(
  Hf,
  S0,
  V0,
  a,
  b,
  c,
  d,
  alpha,
  K0,
  dK0,
  d2K0,
  PGPa,
  TK,
  n,
  Vmax,
  Smax,
  Tc0,
  doLandau
)
}
\description{
Adapted from matlab code by Jesse Walters
}
