% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity.R
\name{Landau}
\alias{Landau}
\title{Landau Model - Holland and Powell 2011}
\usage{
Landau(Tc0, Vmax, Smax, Pbar, TK)
}
\description{
Landau Model - Holland and Powell 2011
}
\details{
the HP2011 database uses a corrected Landau model for some
minerals. This correction was not incorporated into the HP2011 paper;
however, this is code uses the same formulism as Thermocalc. For the
original HP1998 approach, please see the other code. In addition, thermal
expansivity and bulk modulus are not included. In practice, HP set the Vf
term of the volume integral to 1. (JW)
Adapted from matlab code by Jesse Walters
}
