% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activity.R
\name{TEOS}
\alias{TEOS}
\title{THERMOCALC 2011 (Tait) EOS FOR SOLIDS
Adapted from matlab code by Jesse Walters}
\usage{
TEOS(V0, alpha, K0, dK0, d2K0, Pbar, TK, S0, n)
}
\description{
THERMOCALC 2011 (Tait) EOS FOR SOLIDS
Adapted from matlab code by Jesse Walters
}
